+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Fixed usage of *piral.json* in monorepo scenarios for templating new pilets
 - Fixed deferred initialization for standalone Blazor applications using `piral-blazor`
+- Fixed initial layout decision in `piral-page-layouts` (#790)
 - Removed `always-auth` from all generated *.npmrc* files
 - Added support for Angular 21 in `piral-ng`